chore: prepare for product download
