+++ src/RoadRegistry.BackOffice.Api/RoadRegistry.BackOffice.Api.csproj
@@ -22,5 +22,8 @@
   <ItemGroup>
     <Compile Include="..\..\SolutionInfo.cs" Link="Properties\SolutionInfo.cs" />
   </ItemGroup>
+  <ItemGroup>
+    <Folder Include="ZipArchiveWriters\ForProduct" />
+  </ItemGroup>
   <Import Project="..\..\.paket\Paket.Restore.targets" />
 </Project>
