+++ src/RoadRegistry.BackOffice.Api/ZipArchiveWriters/CompositeZipArchiveWriter.cs
@@ -5,17 +5,18 @@ namespace RoadRegistry.BackOffice.Api.ZipArchiveWriters
     using System.Threading;
     using System.Threading.Tasks;
     using Editor.Schema;
+    using Microsoft.EntityFrameworkCore;
 
-    public class CompositeZipArchiveWriter : IZipArchiveWriter
+    public class CompositeZipArchiveWriter<TContext> : IZipArchiveWriter<TContext> where TContext : DbContext
     {
-        private readonly IZipArchiveWriter[] _writers;
+        private readonly IZipArchiveWriter<TContext>[] _writers;
 
-        public CompositeZipArchiveWriter(params IZipArchiveWriter[] writers)
+        public CompositeZipArchiveWriter(params IZipArchiveWriter<TContext>[] writers)
         {
             _writers = writers ?? throw new ArgumentNullException(nameof(writers));
         }
 
-        public async Task WriteAsync(ZipArchive archive, EditorContext context, CancellationToken cancellationToken)
+        public async Task WriteAsync(ZipArchive archive, TContext context, CancellationToken cancellationToken)
         {
             if (archive == null) throw new ArgumentNullException(nameof(archive));
             if (context == null) throw new ArgumentNullException(nameof(context));
