+++ src/RoadRegistry.BackOffice.Api/ZipArchiveWriters/DbaseFileArchiveWriter.cs
@@ -9,8 +9,9 @@ namespace RoadRegistry.BackOffice.Api.ZipArchiveWriters
     using System.Threading.Tasks;
     using Be.Vlaanderen.Basisregisters.Shaperon;
     using Editor.Schema;
+    using Microsoft.EntityFrameworkCore;
 
-    public class DbaseFileArchiveWriter : IZipArchiveWriter
+    public class DbaseFileArchiveWriter<TContext> : IZipArchiveWriter<TContext> where TContext : DbContext
     {
         private readonly string _filename;
         private readonly DbaseSchema _schema;
@@ -25,7 +26,7 @@ namespace RoadRegistry.BackOffice.Api.ZipArchiveWriters
             _encoding = encoding ?? throw new ArgumentNullException(nameof(encoding));
         }
 
-        public async Task WriteAsync(ZipArchive archive, EditorContext context, CancellationToken cancellationToken)
+        public async Task WriteAsync(ZipArchive archive, TContext context, CancellationToken cancellationToken)
         {
             if (archive == null) throw new ArgumentNullException(nameof(archive));
             if (context == null) throw new ArgumentNullException(nameof(context));
