+++ src/RoadRegistry.BackOffice.Api/ZipArchiveWriters/ForEditor/GradeSeparatedJunctionArchiveWriter.cs
@@ -1,4 +1,4 @@
-namespace RoadRegistry.BackOffice.Api.ZipArchiveWriters
+namespace RoadRegistry.BackOffice.Api.ZipArchiveWriters.ForEditor
 {
     using System;
     using System.IO;
@@ -8,18 +8,17 @@ namespace RoadRegistry.BackOffice.Api.ZipArchiveWriters
     using System.Threading;
     using System.Threading.Tasks;
     using Be.Vlaanderen.Basisregisters.Shaperon;
-    using Core;
     using Editor.Schema;
-    using Editor.Schema.Organizations;
+    using Editor.Schema.GradeSeparatedJunctions;
     using Microsoft.EntityFrameworkCore;
     using Microsoft.IO;
 
-    public class OrganizationsToZipArchiveWriter : IZipArchiveWriter
+    public class GradeSeparatedJunctionArchiveWriter : IZipArchiveWriter<EditorContext>
     {
         private readonly RecyclableMemoryStreamManager _manager;
         private readonly Encoding _encoding;
 
-        public OrganizationsToZipArchiveWriter(RecyclableMemoryStreamManager manager, Encoding encoding)
+        public GradeSeparatedJunctionArchiveWriter(RecyclableMemoryStreamManager manager, Encoding encoding)
         {
             _manager = manager ?? throw new ArgumentNullException(nameof(manager));
             _encoding = encoding ?? throw new ArgumentNullException(nameof(encoding));
@@ -30,12 +29,13 @@ namespace RoadRegistry.BackOffice.Api.ZipArchiveWriters
             if (archive == null) throw new ArgumentNullException(nameof(archive));
             if (context == null) throw new ArgumentNullException(nameof(context));
 
-            var dbfEntry = archive.CreateEntry("LstOrg.dbf");
+            var count = await context.GradeSeparatedJunctions.CountAsync(cancellationToken);
+            var dbfEntry = archive.CreateEntry("RltOgkruising.dbf");
             var dbfHeader = new DbaseFileHeader(
                 DateTime.Now,
                 DbaseCodePage.Western_European_ANSI,
-                new DbaseRecordCount(await context.Organizations.CountAsync(cancellationToken) + Organization.PredefinedTranslations.All.Length),
-                OrganizationDbaseRecord.Schema
+                new DbaseRecordCount(count),
+                GradeSeparatedJunctionDbaseRecord.Schema
             );
             using (var dbfEntryStream = dbfEntry.Open())
             using (var dbfWriter =
@@ -43,19 +43,13 @@ namespace RoadRegistry.BackOffice.Api.ZipArchiveWriters
                     dbfHeader,
                     new BinaryWriter(dbfEntryStream, _encoding, true)))
             {
-                var dbfRecord = new OrganizationDbaseRecord();
-                foreach (var predefined in Organization.PredefinedTranslations.All)
-                {
-                    dbfRecord.ORG.Value = predefined.Identifier;
-                    dbfRecord.LBLORG.Value = predefined.Name;
-                    dbfWriter.Write(dbfRecord);
-                }
-
-                foreach (var data in context.Organizations.OrderBy(_ => _.SortableCode).Select(_ => _.DbaseRecord))
+                var dbfRecord = new GradeSeparatedJunctionDbaseRecord();
+                foreach (var data in context.GradeSeparatedJunctions.OrderBy(_ => _.Id).Select(_ => _.DbaseRecord))
                 {
                     dbfRecord.FromBytes(data, _manager, _encoding);
                     dbfWriter.Write(dbfRecord);
                 }
+
                 dbfWriter.Writer.Flush();
                 await dbfEntryStream.FlushAsync(cancellationToken);
             }
