+++ src/RoadRegistry.BackOffice.Api/ZipArchiveWriters/ForEditor/RoadSegmentNationalRoadAttributesToZipArchiveWriter.cs
@@ -1,4 +1,4 @@
-namespace RoadRegistry.BackOffice.Api.ZipArchiveWriters
+namespace RoadRegistry.BackOffice.Api.ZipArchiveWriters.ForEditor
 {
     using System;
     using System.IO;
@@ -13,12 +13,12 @@ namespace RoadRegistry.BackOffice.Api.ZipArchiveWriters
     using Microsoft.EntityFrameworkCore;
     using Microsoft.IO;
 
-    public class RoadSegmentEuropeanRoadAttributesToZipArchiveWriter : IZipArchiveWriter
+    public class RoadSegmentNationalRoadAttributesToZipArchiveWriter : IZipArchiveWriter<EditorContext>
     {
         private readonly RecyclableMemoryStreamManager _manager;
         private readonly Encoding _encoding;
 
-        public RoadSegmentEuropeanRoadAttributesToZipArchiveWriter(RecyclableMemoryStreamManager manager,
+        public RoadSegmentNationalRoadAttributesToZipArchiveWriter(RecyclableMemoryStreamManager manager,
             Encoding encoding)
         {
             _manager = manager ?? throw new ArgumentNullException(nameof(manager));
@@ -30,13 +30,13 @@ namespace RoadRegistry.BackOffice.Api.ZipArchiveWriters
             if (archive == null) throw new ArgumentNullException(nameof(archive));
             if (context == null) throw new ArgumentNullException(nameof(context));
 
-            var count = await context.RoadSegmentEuropeanRoadAttributes.CountAsync(cancellationToken);
-            var dbfEntry = archive.CreateEntry("AttEuropweg.dbf");
+            var count = await context.RoadSegmentNationalRoadAttributes.CountAsync(cancellationToken);
+            var dbfEntry = archive.CreateEntry("AttNationweg.dbf");
             var dbfHeader = new DbaseFileHeader(
                 DateTime.Now,
                 DbaseCodePage.Western_European_ANSI,
                 new DbaseRecordCount(count),
-                RoadSegmentEuropeanRoadAttributeDbaseRecord.Schema
+                RoadSegmentNationalRoadAttributeDbaseRecord.Schema
             );
             using (var dbfEntryStream = dbfEntry.Open())
             using (var dbfWriter =
@@ -44,8 +44,8 @@ namespace RoadRegistry.BackOffice.Api.ZipArchiveWriters
                     dbfHeader,
                     new BinaryWriter(dbfEntryStream, _encoding, true)))
             {
-                var dbfRecord = new RoadSegmentEuropeanRoadAttributeDbaseRecord();
-                foreach (var data in context.RoadSegmentEuropeanRoadAttributes.OrderBy(_ => _.Id).Select(_ => _.DbaseRecord))
+                var dbfRecord = new RoadSegmentNationalRoadAttributeDbaseRecord();
+                foreach (var data in context.RoadSegmentNationalRoadAttributes.OrderBy(_ => _.Id).Select(_ => _.DbaseRecord))
                 {
                     dbfRecord.FromBytes(data, _manager, _encoding);
                     dbfWriter.Write(dbfRecord);
