+++ src/RoadRegistry.BackOffice.Api/ZipArchiveWriters/ForEditor/RoadSegmentNumberedRoadAttributesToZipArchiveWriter.cs
@@ -1,4 +1,4 @@
-namespace RoadRegistry.BackOffice.Api.ZipArchiveWriters
+namespace RoadRegistry.BackOffice.Api.ZipArchiveWriters.ForEditor
 {
     using System;
     using System.IO;
@@ -9,16 +9,17 @@ namespace RoadRegistry.BackOffice.Api.ZipArchiveWriters
     using System.Threading.Tasks;
     using Be.Vlaanderen.Basisregisters.Shaperon;
     using Editor.Schema;
-    using Editor.Schema.GradeSeparatedJunctions;
+    using Editor.Schema.RoadSegments;
     using Microsoft.EntityFrameworkCore;
     using Microsoft.IO;
 
-    public class GradeSeparatedJunctionArchiveWriter : IZipArchiveWriter
+    public class RoadSegmentNumberedRoadAttributesToZipArchiveWriter : IZipArchiveWriter<EditorContext>
     {
         private readonly RecyclableMemoryStreamManager _manager;
         private readonly Encoding _encoding;
 
-        public GradeSeparatedJunctionArchiveWriter(RecyclableMemoryStreamManager manager, Encoding encoding)
+        public RoadSegmentNumberedRoadAttributesToZipArchiveWriter(RecyclableMemoryStreamManager manager,
+            Encoding encoding)
         {
             _manager = manager ?? throw new ArgumentNullException(nameof(manager));
             _encoding = encoding ?? throw new ArgumentNullException(nameof(encoding));
@@ -29,13 +30,13 @@ namespace RoadRegistry.BackOffice.Api.ZipArchiveWriters
             if (archive == null) throw new ArgumentNullException(nameof(archive));
             if (context == null) throw new ArgumentNullException(nameof(context));
 
-            var count = await context.GradeSeparatedJunctions.CountAsync(cancellationToken);
-            var dbfEntry = archive.CreateEntry("RltOgkruising.dbf");
+            var count = await context.RoadSegmentNumberedRoadAttributes.CountAsync(cancellationToken);
+            var dbfEntry = archive.CreateEntry("AttGenumweg.dbf");
             var dbfHeader = new DbaseFileHeader(
                 DateTime.Now,
                 DbaseCodePage.Western_European_ANSI,
                 new DbaseRecordCount(count),
-                GradeSeparatedJunctionDbaseRecord.Schema
+                RoadSegmentNumberedRoadAttributeDbaseRecord.Schema
             );
             using (var dbfEntryStream = dbfEntry.Open())
             using (var dbfWriter =
@@ -43,8 +44,8 @@ namespace RoadRegistry.BackOffice.Api.ZipArchiveWriters
                     dbfHeader,
                     new BinaryWriter(dbfEntryStream, _encoding, true)))
             {
-                var dbfRecord = new GradeSeparatedJunctionDbaseRecord();
-                foreach (var data in context.GradeSeparatedJunctions.OrderBy(_ => _.Id).Select(_ => _.DbaseRecord))
+                var dbfRecord = new RoadSegmentNumberedRoadAttributeDbaseRecord();
+                foreach (var data in context.RoadSegmentNumberedRoadAttributes.OrderBy(_ => _.Id).Select(_ => _.DbaseRecord))
                 {
                     dbfRecord.FromBytes(data, _manager, _encoding);
                     dbfWriter.Write(dbfRecord);
