+++ src/RoadRegistry.BackOffice.Api/ZipArchiveWriters/ReadCommittedZipArchiveWriter.cs
@@ -8,16 +8,16 @@ namespace RoadRegistry.BackOffice.Api.ZipArchiveWriters
     using Editor.Schema;
     using Microsoft.EntityFrameworkCore;
 
-    public class ReadCommittedZipArchiveWriter : IZipArchiveWriter
+    public class ReadCommittedZipArchiveWriter<TContext> : IZipArchiveWriter<TContext> where TContext: DbContext
     {
-        private readonly IZipArchiveWriter _writer;
+        private readonly IZipArchiveWriter<TContext> _writer;
 
-        public ReadCommittedZipArchiveWriter(IZipArchiveWriter writer)
+        public ReadCommittedZipArchiveWriter(IZipArchiveWriter<TContext> writer)
         {
             _writer = writer ?? throw new ArgumentNullException(nameof(writer));
         }
 
-        public async Task WriteAsync(ZipArchive archive, EditorContext context, CancellationToken cancellationToken)
+        public async Task WriteAsync(ZipArchive archive, TContext context, CancellationToken cancellationToken)
         {
             if (archive == null) throw new ArgumentNullException(nameof(archive));
             if (context == null) throw new ArgumentNullException(nameof(context));
