+++ src/RoadRegistry.BackOffice.UI/src/Download.elm
@@ -39,7 +39,7 @@ init url =
     ( { header = Header.init |> Header.downloadBecameActive
       , download =
             { title = "Register dump"
-            , url = String.concat [ url, "/v1/download" ]
+            , url = String.concat [ url, "/v1/download/for-editor" ]
             , downloading = False
             , progressing = False
             , progress = ""
